requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2019.9.11
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
dropbox==11.0.0
idna==2.10
multidict==4.7.6
pipenv==2018.11.26
ply==3.11
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
stone==3.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==16.7.7
virtualenv-clone==0.5.3
yarl==1.5.1
